requirements.txt
pandas==1.5.3
openpyxl==3.0.10
python-dotenv==1.1.0
pytest==6.2.5